refactor: move goUp and changeDirrectory functions
