+++ src/handle.js
@@ -1,6 +1,4 @@
-import fs from 'node:fs';
-import path from 'node:path';
-import { rootDir } from './directory.js';
+import { goUp, changeDirectory } from './directory.js';
 
 export const handleCommand = (input) => {
     const [command, ...args] = input.split(' ');
@@ -22,36 +20,5 @@ export const handleCommand = (input) => {
 
         default:
         console.error('Invalid input');
-    }
-};
-
-const changeDirectory = (input) => {
-    const newPath = path.isAbsolute(input) ? input : path.join(process.cwd(), input);
-    try {
-        if (fs.existsSync(newPath) && fs.statSync(newPath).isDirectory() && newPath.startsWith(rootDir)) {
-            process.chdir(newPath);
-        } else {
-            throw new Error();
-        }
-    } catch {
-        console.log('Operation failed');
-    }
-};
-
-export const goUp = () => {
-    const parentDir = path.resolve(process.cwd(), '..');
-
-    try {
-        if (
-            parentDir.startsWith(rootDir) 
-            && fs.statSync(parentDir).isDirectory 
-            && fs.existsSync(parentDir)
-        ) {
-            process.chdir(parentDir);
-        } else {
-            throw new Error();
-        }
-    } catch {
-        console.log('Operation failed');
     }
 };
